Heart_Disease_Project/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
streamlit
joblib
ngrok